AI-Summarizer/requirements.txt
streamlit
langchain
transformers
torch
pillow
python-docx
python-pptx
pymupdf
google-auth
google-auth-oauthlib
google-auth-httplib2
PyPDF2
pdfplumber
numpy
SentencePiece
langchain-community
torchvision
torchaudio